pyproject.toml
[project]
name = "gaiia-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "pydantic-settings>=2.11.0",
    "tavily-python>=0.7.12",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "ruff>=0.14.1",
]
